rename to Pomelo.EntityFrameworkCore.MySql.PerfTests
