+++ test/Pomelo.EntityFrameworkCore.MySql.PerfTests/Models/DataTypes.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 
-namespace Pomelo.EntityFrameworkCore.MySql.Functional.Models
+namespace Pomelo.EntityFrameworkCore.MySql.PerfTests.Models
 {
     public class DataTypes
     {
