+++ test/Pomelo.EntityFrameworkCore.MySql.PerfTests/Models/Post.cs
@@ -1,4 +1,4 @@
-﻿namespace Pomelo.EntityFrameworkCore.MySql.Functional.Models
+﻿namespace Pomelo.EntityFrameworkCore.MySql.PerfTests.Models
 {
     public class Post
     {
