+++ test/Pomelo.EntityFrameworkCore.MySql.PerfTests/README.md
