+++ test/Pomelo.EntityFrameworkCore.MySql.PerfTests/project.json
