+++ test/Pomelo.EntityFrameworkCore.MySql.PerfTests/scripts/rebuild.ps1
