+++ test/Pomelo.EntityFrameworkCore.MySql.PerfTests/scripts/rebuild.sh
