+++ test/Pomelo.EntityFrameworkCore.MySql.PerfTests/scripts/stress.sh
